requirements.txt
# local package
-e .
numpy
jupyter
matplotlib
torch

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pytest
yapf
pytorch-lightning
sklearn
optuna 
coloredlogs
seaborn
networkx